push fix bug

+++ database/migrations/2024_08_03_193501_create_guide_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('guides', function (Blueprint $table) {
             $table->id('id');
             $table->string('nama_guide', 100);
-            $table->decimal('salary', 10, 2);
+            $table->integer('salary'); // Menggunakan integer tanpa desimal
             $table->unsignedBigInteger('kriteria_id')->nullable(); //tidak usah di isi pada saat create tapi ditampilkan saja kriteria nya
             $table->text('deskripsi_guide')->nullable();
             $table->string('nomer_hp', 20);
